Return ByteSize from HttpResponseHeaders:::GetContentLength()
ByteCount is deprecated.

ByteCount::InBytes() returns int64_t, and ByteSize::InBytes() returns
values in the range 0...max(int64_t), so call sites will not overflow
unless the ByteCount usage was already unsafe.

Also fixes a variable name in HttpResponseHeaders::GetInt64HeaderValue().

Bug: 448661443
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8065298
Commit-Queue: Avi Drissman <avi@chromium.org>
Auto-Submit: Joe Mason <joenotcharles@google.com>
Owners-Override: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1660257}

+++ content/browser/blob_storage/blob_url_unittest.cc
@@ -9,6 +9,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include <memory>
 #include <string_view>
 
+#include "base/byte_size.h"
 #include "base/compiler_specific.h"
 #include "base/containers/span.h"
 #include "base/files/file_path.h"
@@ -192,8 +193,9 @@ class BlobURLTest : public testing::Test {
     expected_status_code_ = 200;
     expected_response_ = expected_response;
     TestRequest("GET", net::HttpRequestHeaders());
-    EXPECT_EQ(expected_content_length,
-              response_headers_->GetContentLength()->InBytes());
+    EXPECT_EQ(
+        expected_content_length,
+        static_cast<int64_t>(response_headers_->GetContentLength()->InBytes()));
   }
 
   void TestErrorRequest(int expected_error_code) {
@@ -491,7 +493,7 @@ TEST_F(BlobURLTest, TestGetRangeRequest1) {
   expected_response_ = result.substr(5, 10 - 5 + 1);
   TestRequest("GET", extra_headers);
 
-  EXPECT_EQ(6, response_headers_->GetContentLength()->InBytes());
+  EXPECT_EQ(6u, response_headers_->GetContentLength()->InBytes());
   EXPECT_FALSE(response_metadata_.has_value());
 
   int64_t first = 0, last = 0, length = 0;
@@ -512,7 +514,7 @@ TEST_F(BlobURLTest, TestGetRangeRequest2) {
   expected_response_ = result.substr(result.length() - 10);
   TestRequest("GET", extra_headers);
 
-  EXPECT_EQ(10, response_headers_->GetContentLength()->InBytes());
+  EXPECT_EQ(10u, response_headers_->GetContentLength()->InBytes());
   EXPECT_FALSE(response_metadata_.has_value());
 
   int64_t total = GetTotalBlobLength();
@@ -534,7 +536,7 @@ TEST_F(BlobURLTest, TestGetRangeRequest3) {
   expected_response_ = result.substr(0, 3);
   TestRequest("GET", extra_headers);
 
-  EXPECT_EQ(3, response_headers_->GetContentLength()->InBytes());
+  EXPECT_EQ(3u, response_headers_->GetContentLength()->InBytes());
   EXPECT_FALSE(response_metadata_.has_value());
 
   int64_t first = 0, last = 0, length = 0;
@@ -570,7 +572,7 @@ TEST_F(BlobURLTest, TestSideData) {
   expected_status_code_ = 200;
   expected_response_ = kTestDataHandleData2;
   TestRequest("GET", net::HttpRequestHeaders());
-  EXPECT_EQ(static_cast<int>(std::size(kTestDataHandleData2) - 1),
+  EXPECT_EQ(std::size(kTestDataHandleData2) - 1,
             response_headers_->GetContentLength()->InBytes());
 
   EXPECT_EQ(std::string(kTestDiskCacheSideData), *response_metadata_);
@@ -583,7 +585,7 @@ TEST_F(BlobURLTest, TestZeroSizeSideData) {
   expected_status_code_ = 200;
   expected_response_ = kTestDataHandleData2;
   TestRequest("GET", net::HttpRequestHeaders());
-  EXPECT_EQ(static_cast<int>(std::size(kTestDataHandleData2) - 1),
+  EXPECT_EQ(std::size(kTestDataHandleData2) - 1,
             response_headers_->GetContentLength()->InBytes());
 
   EXPECT_FALSE(response_metadata_.has_value());
